### 2025-04-23 #8097 @editorialbot
[PRE REVIEW]: PathSim - A System Simulation Framework
**Submitting author:** @milanofthe (<a href="http://orcid.org/0009-0006-5964-6115">Milan Rother</a>)
**Repository:** https://github.com/milanofthe/pathsim
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.6.6
