### 2025-04-23 #8097 @milanofthe
Thanks @danielskatz! Will do! Regarding published research, its still quite early for the package, but I know of some applications in the domain of chemical processes for fusion reactors.
