### 2025-04-23 588d93f
fixed capital P for Python in paper.bib, added country to affiliation in paper.md
