+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 @misc{bdsimRepo,
   author       = {Peter Corke},
-  title        = {bdsim: Simulate dynamic systems expressed in block diagram form using Python},
+  title        = {bdsim: Simulate dynamic systems expressed in block diagram form using {P}ython},
   howpublished = {\url{https://github.com/petercorke/bdsim}},
   month        = {May},
   year         = {2020},
@@ -34,7 +34,7 @@
   number    = {17}, 
   pages     = {396}, 
   author    = {Benjamin W. l. Margolis}, 
-  title     = {SimuPy: A Python framework for modeling and simulating dynamical systems}, 
+  title     = {SimuPy: A {P}ython framework for modeling and simulating dynamical systems}, 
   journal   = {Journal of Open Source Software} 
   }
 
@@ -65,7 +65,7 @@
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
   title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+            Computing in {P}ython}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -82,7 +82,7 @@
   Volume    = {9},
   Number    = {3},
   Pages     = {90--95},
-  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  abstract  = {Matplotlib is a 2D graphics package used for {P}ython for
   application development, interactive scripting, and publication-quality
   image generation across user interfaces and operating systems.},
   publisher = {IEEE COMPUTER SOC},
@@ -118,7 +118,7 @@
 @INPROCEEDINGS{pythoncontrol,
   author={Fuller, Sawyer and Greiner, Ben and Moore, Jason and Murray, Richard and van Paassen, René and Yorke, Rory},
   booktitle={2021 60th IEEE Conference on Decision and Control (CDC)}, 
-  title={The Python Control Systems Library (python-control)}, 
+  title={The {P}ython {C}ontrol {S}ystems {L}ibrary (python-control)}, 
   year={2021},
   volume={},
   number={},
