### 2025-04-23 #8097 @milanofthe
@danielskatz thanks for your comments again, I made the following changes:
- Added the country to my affiliation in `paper.md`.
- Added more specificity to the **Statement of Need** section, its tricky to be super specific here, due to the generality of the package.
- Fixed the capitalization of Python in `paper.bib`.

I know of users who use it to simulate radar systems, communication systems, fusion reactor fuel cycles, HiL environments for thermal systems. The goal of this package is to be as flexible and general as possible, while having super robust solvers for all kinds of systems. Its one of the great challenges, that users can throw anything at it and it should still be able to handle it. If you think it makes sense to be even more specific in the **Statement of Need**, I could add these examples.
