### 2025-04-23 #8097 @milanofthe
Id like to suggest @mbudisic because of his background in dynamical systems and numerics for differential equations (ode solvers are at the core of PathSim), @lukashof because of his background in process simulation (while not strictly a process simulator, processes can be simulated with PathSim), @thermalogic because his background in system simulation (PathSim is a system simulation framework), @sea-bass because of his background in robotics and control systems and @t-gross because of her background in flow simulation (chemical process flows are one area of application). I think they might even find this interesting.
