### 2025-04-24 005245b
added table to Adder block for IO
