### 2025-04-24 684d18f
better docstrings
