### 2025-04-24 c7fc84d
better docstrings
