+++ src/pathsim/blocks/adder.py
@@ -14,6 +14,7 @@ import numpy as np
 from ._block import Block
 
 from ..utils.utils import dict_to_array
+
 from ..optim.operator import Operator
 
 
