### 2025-04-24 ab22bb7
added Register class to replace inputs and outputs dicts for blocks
