### 2025-04-24 44ba356
refactor Switch to use Register
