### 2025-04-24 e470a73
refactor Scope and RealtimeScope to use Register
