### 2025-04-24 df0645e
refactor Spectrum and RealtimeSpectrum to use Register
