### 2025-04-24 747f5b9
refactor Multiplier to use Register
