### 2025-04-24 d260398
refactor StateSpace to use Register
