### 2025-04-24 aea1232
refactor Integrator to use Register
