+++ src/pathsim/blocks/integrator.py
@@ -108,7 +108,7 @@ class Integrator(Block):
         error : float
             deviation to previous iteration for convergence control
         """
-        self.outputs = array_to_dict(self.engine.get())
+        self.outputs.update_from_array(self.engine.get())
         return 0.0
 
 
@@ -127,7 +127,7 @@ class Integrator(Block):
         error : float
             solver residual norm
         """
-        f = dict_to_array(self.inputs)
+        f = self.inputs.to_array()
         return self.engine.solve(f, None, dt)
 
 
@@ -150,5 +150,5 @@ class Integrator(Block):
         scale : float
             timestep rescale from adaptive integrators
         """
-        f = dict_to_array(self.inputs)
+        f = self.inputs.to_array()
         return self.engine.step(f, dt)
