### 2025-04-24 43b8ef8
refactor Function to use Register
