### 2025-04-24 7de948c
refactor Delay to use Register
