+++ src/pathsim/blocks/ctrl.py
@@ -145,8 +145,8 @@ class PID(Block):
             absolute error to previous iteration for convergence control
         """
         x, u = self.engine.get(), self.inputs[0]
-        _out, self.outputs[0] = self.outputs[0], self.op_alg(x, u, t)
-        return abs(_out - self.outputs[0])
+        y = self.op_alg(x, u, t)
+        return self.outputs.update_from_array_max_err(y)
 
 
     def solve(self, t, dt):
