### 2025-04-24 a313553
refactor adder to use Register
