+++ src/pathsim/blocks/adder.py
@@ -132,6 +132,6 @@ class Adder(Block):
         error : float
             absolute error to previous iteration for convergence control
         """
-        u = dict_to_array(self.inputs)
-        _out, self.outputs[0] = self.outputs[0], self.op_alg(u)
-        return abs(_out - self.outputs[0])
+        u = self.inputs.to_array()
+        y = self.op_alg(u)
+        return self.outputs.update_from_array_max_err(y)
