### 2025-04-24 64500cf
Register.to_array now uses list comprehension, and fixed typo in Scope.sample
