### 2025-04-24 0f5a502
remove Register._to_num helper
