### 2025-04-24 752c731
reimplement helper method
