### 2025-04-25 f3d0556
refactor ADC and DAC to work with Register
