### 2025-04-25 fb48e5e
refactor noise sources in blocks/rf/noise to work with Register and improved docstrings
