### 2025-04-25 3ecaaf3
removed use of np.atleast_1d from Register to improve performance, instead special case for np.isscalar
