### 2025-04-25 10bfdb2
removing Register._get_num helper, using np.flatten instead for performance
