### 2025-04-25 f1a4c4a
fixed input register access for DAC and ADC blocks
