### 2025-04-25 2685e5a
refactor Subsystem to work with Register
