### 2025-04-25 4e74d4b
added docstrings to Register
