### 2025-04-25 e8e6dcb
updated header in register.py
