+++ src/pathsim/utils/register.py
@@ -1,6 +1,9 @@
 #########################################################################################
 ##
 ##                                   Register Class
+##                            (pathsim/utils/register.py)
+##
+##                                 Milan Rother 2025
 ##
 #########################################################################################
 
