+++ docs/source/modules/pathsim.utils.register.rst
@@ -1,0 +1,7 @@
+pathsim.utils.register module
+==========================
+
+.. automodule:: pathsim.utils.register
+   :members:
+   :show-inheritance:
+   :undoc-members:
