### 2025-04-25 0bc0970
cleanup of unused methods from utils.utils
