### 2025-04-25 9fbbea4
cleanup of unused functions, primarily helpers in utils/utils.py
