+++ src/pathsim/utils/utils.py
@@ -9,197 +9,7 @@
 
 # IMPORTS ==============================================================================
 
-import numpy as np
-
-from .. _constants import TOLERANCE
-
-
-# HELPERS FOR SIMULATION ===============================================================
-
-def dict_to_array(a):
-    """convert a dict with integer keys to a numpy array
-
-    Parameters
-    ----------
-    a : dict[int: int, float, complex]
-        dict to convert to numpy array
-        
-    Returns
-    -------
-    out : array[int, float, complex]
-        converted array
-    """
-    return np.array(
-        [a[k].item() if hasattr(a[k], "item") else a[k] 
-            for k in sorted(a.keys())]
-        )
-
-
-def array_to_dict(a):
-    """convert a numpy array to a dict with integer keys
-
-    Parameters
-    ----------
-    a : array[int, float, complex]
-        numpy array to convert
-    
-    Returns
-    -------
-    out: dict[int: int, float, complex]
-        converted dict
-    """
-    if np.isscalar(a): return {0:a}
-    else: return dict(enumerate(a))
-
-
-def rel_error(a, b):
-    """Computes the relative error between two scalars.
-    
-    It is robust to one of them being close to zero and 
-    falls back to the absolute error in this case.
-
-    Notes
-    ----- 
-    this is actually faster then inlining the 
-    branching into the return statement
-    
-    Parameters
-    ----------
-    a : float, int, complex
-        first number
-    b : float, int, complex
-        second number
-    
-    Returns
-    -------
-    err : float
-        retative error
-    """
-    if abs(a) < TOLERANCE: return abs(b)
-    else: return abs((a - b)/a)
-
-
-def abs_error(a, b):
-    """Computes the absolute error between two scalars.
-
-    Parameters
-    ----------
-    a : float, int, complex
-        first number
-    b : float, int, complex
-        second number
-    
-    Returns
-    -------
-    err : float
-        absolute error
-    """
-    return abs(a - b)
-
-
-def max_error(a, b):
-    """Computes the maximum absolute error / deviation between two
-    iterables such as lists with numerical values. Returns a scalar 
-    value representing the maximum deviation.
-
-    Notes
-    -----
-    this is actually faster then 'max' over a list comprehension
-    
-    Parameters
-    ----------
-    a : iterable[float, int, complex]
-        first iterable with numerical values
-    b : iterable[float, int, complex]
-        second iterable with numerical values
-    
-    Returns
-    -------
-    err : float
-        maximum absolute error
-    """
-    max_err = 0.0
-    for err in map(abs_error, a, b):
-        if err > max_err: 
-            max_err = err
-    return max_err
-
-
-def max_rel_error(a, b):
-    """Computes the maximum relative error between two iterables 
-    such as lists with numerical values. 
-
-    It is robust to one of them being zero and falls back to the 
-    absolute error in this case. 
-
-    It returns a scalar value representing the maximum relative error. 
-
-    Notes
-    -----
-    this is actually faster then 'max' over a list comprehension
-
-    Parameters
-    ----------
-    a : iterable[float, int, complex]
-        first iterable with numerical values
-    b : iterable[float, int, complex]
-        second iterable with numerical values
-    
-    Returns
-    -------
-    err : float
-        maximum retative error
-    """
-    max_err = 0.0
-    for err in map(rel_error, a, b):
-        if err > max_err: 
-            max_err = err
-    return max_err
-
-
-def max_error_dicts(a, b):
-    """Computes the maximum absolute error between two dictionaries 
-    with numerical values. 
-
-    It returns a scalar value representing the maximum absolute error.
-
-    Parameters
-    ----------
-    a : dict[int: float, int, complex]
-        first dict with numerical values
-    b : dict[int: float, int, complex]
[…]
