### 2025-04-25 3aad263
added test_register to test suite
