### 2025-04-25 3bcb94a
adding tests for Register
