+++ tests/pathsim/utils/test_register.py
@@ -12,25 +12,227 @@
 import unittest
 import numpy as np
 
+from collections import defaultdict
+
 from pathsim.utils.register import Register
 
 
 # TESTS ================================================================================
 
-class TestPortReference(unittest.TestCase):
+class TestRegister(unittest.TestCase):
     """
     test the 'Register' class
     """
 
-    def test_init(self): pass
-    def test_len(self): pass
-    def test_iter(self): pass
-    def test_reset(self): pass
-    def test_to_array(self): pass
-    def test_update_from_array(self): pass
-    def test_update_from_array_max_err(self): pass
-    def test_setitem(self): pass
-    def test_getitem(self): pass
+    def test_init(self):
+        #default
+        R = Register()
+        self.assertTrue(isinstance(R._values, defaultdict))
+        self.assertEqual(len(R._values), 1)
+        self.assertEqual(R[0], 0.0)
+
+        #specific
+        R = Register(size=20)
+        self.assertTrue(isinstance(R._values, defaultdict))
+        self.assertEqual(len(R._values), 20)
+        self.assertEqual(R[0], 0.0)
+        self.assertEqual(R[19], 0.0)
+
+        #accessing a key beyond initial size uses defaultdict factory
+        self.assertEqual(R[20], 0.0)
+        
+        #but the lengt increases 
+        self.assertEqual(len(R), 21)
+
+
+    def test_len(self):
+        #default
+        R = Register()
+        self.assertEqual(len(R), 1)
+
+        #specific
+        R = Register(size=5)
+        self.assertEqual(len(R), 5)
+
+        # After adding items
+        R[5] = 10.0
+        R[10] = 20.0
+        self.assertEqual(len(R), 7) # 0, 1, 2, 3, 4, 5, 10
+
+
+    def test_iter(self):
+        #specific
+        R = Register(3)
+        R[0] = 1.0
+        R[1] = 2.0
+        R[2] = 3.0
+
+        #add out of order
+        R[5] = 5.0
+        R[4] = 4.0
+        expected = [1.0, 2.0, 3.0, 4.0, 5.0]
+
+        #tests sorting
+        self.assertEqual([v for v in R], expected) 
+        self.assertEqual(len([v for v in R]), 5)
+
+
+    def test_reset(self):
+        
+        R = Register(3)
+        
+        R[0] = 1.0
+        R[1] = -2.5
+        R[2] = 100.0
+        R[5] = 99.0 
+
+        self.assertEqual(len(R), 4)
+        
+        R.reset()
+
+        #length remains same
+        self.assertEqual(len(R), 4) 
+        
+        self.assertEqual(R[0], 0.0)
+        self.assertEqual(R[1], 0.0)
+        self.assertEqual(R[2], 0.0)
+        
+        #reset dynamic key too
+        self.assertEqual(R[5], 0.0) 
+
+        #default value for non-existent key after reset
+        self.assertEqual(R[10], 0.0)
+
+
+    def test_to_array(self):
+        R = Register(3)
+        R[0] = 1.1
+        R[1] = 2.2
+        R[2] = 3.3
+
+        np.testing.assert_array_equal(
+            R.to_array(), 
+            np.array([1.1, 2.2, 3.3])
+            )
+
+        #test with dynamic keys added out of order
+        R[5] = 5.5
+        R[4] = 4.4
+
+        np.testing.assert_array_equal(
+            R.to_array(), 
+            np.array([1.1, 2.2, 3.3, 4.4, 5.5])
+            )
+
+        #test empty initialized (should be size 1 default)
+        R = Register()
+
+        np.testing.assert_array_equal(
[…]
