### 2025-04-26 9271f34
refactor Register for performance, adapted tests, since defaultdict is no longer used
