### 2025-04-26 2985174
fixed docstrings of Register
