### 2025-04-26 3c3273f
more descriptive docstrings for Register
