+++ ROADMAP.md
@@ -1,19 +1,18 @@
 # PathSim Development Roadmap
 
-This is a preliminary roadmap for PathSim's development. To be fair, its more a collection of ideas and potential directions then a real roadmap.
+This is a preliminary roadmap for PathSim's development. To be fair, its more a collection of ideas and potential directions then a real roadmap. 
 
+
+# Algorithms
 
 ## Solvers
 
-- multistep methods (BDF, GEAR) need startup methods to build history
-
+- multistep methods (BDF, GEAR) need startup methods to build history, to maintain consistency order globally
 
 ## Analysis
 
 - periodic steady state solver, probably shooting method with automatic frequency/period detection
 - small signal frequency domain analysis based on linearized system
-- time constant extraction via eigenvalues of linearized dynamic blocks
-
 
 ## Performance
 
@@ -24,19 +23,37 @@ This is a preliminary roadmap for PathSim's development. To be fair, its more a 
 - jit compilation of internal functions in the operators (`Operator`, `DynamicOperator`)
 
 
+# Usability
+
+## Documentation
+
+- the integrators (`Solver` child classes) would benefit from more descriptive docstrings for the api reference including explanations and references (literature) to the specific method used and also some usage suggestions to make solver choice easier and more transparent for the user
+- add more tutorials to the docs, that focus on:
+	- types of analyses and visualization methods (transient -> `Scope`, frequency -> `Spectrum`, steadystate -> ??)
+	- types of available solvers in the PathSim solver suite, *find your own solver*
+	- the block diagram modelling paradigm in general
+	- hierarchical modeling with the `Subsystem` class
+- type hints for everything
+
 ## API
 
-- separate transient and steady state analysis results for scope and spectrum
-- add options to integrators to specify the type of optimizer to be used to solve the implicit update equation, currenly the hybrid `NewtonAnderson` is used, but more flexibility might be nice in the future
+- separate the different kinds of analyses more clearly, transient and steady state analysis results should be separated for `Scope` and `Spectrum` blocks to feel more natural
+- add options to integrators (`ImplicitSolver`) to specify the type of optimizer to be used to solve the implicit update equation, currenly the hybrid `NewtonAnderson` is used, but more flexibility might be nice in the future
+
+## Cross Compatibility
+
+- support for FMI / FMU, import and export of PathSim models and blocks 
+- support for electrical circuits, SPICE netlists 
+- support for s-parameters (touchstone files) 
 
 
-## Testing
+# Testing
 
-- complete testing for all the blocks (currently testing for blocks is mostly top down) including linearization
+## Blocks
+- complete testing for blocks in `pathsim/blocks/rf` and `pathsim/blocks/mixed` (currently testing for these blocks is mostly top down)
+
+## Features
 - test automatic differentiation through `Value` with every block and with linearization
 
-
-## Integration
-
-- support for Modelica / FMI models, import as special blocks
-- special circuit blocks, generated from SPICE netlists
+## Robustness
+- permutation testing for all kinds of system topologies, blocks and solvers
