### 2025-04-27 53c8f74
renaming of _events -> _detected_events in Simulation
