### 2025-04-27 759136b
major refactor of Simulation, separating adaptive/fixed and explicit/implicit solver timestepping loops more clearly
