+++ src/pathsim/simulation.py
@@ -791,6 +791,11 @@ class Simulation:
         return sorted(detected_events, key=lambda e: e[-1])
 
 
+    def _estimated_event_ratio(self, time):
+
+        pass
+
+
     # solving system equations ----------------------------------------------------
 
     def _update(self, t):
@@ -960,7 +965,7 @@ class Simulation:
         self._set_solver(_solver)
 
 
-    # timestepping ----------------------------------------------------------------
+    # timestepping helpers --------------------------------------------------------
 
     def _revert(self):
         """Revert simulation state to previous timestep for adaptive solvers 
@@ -1076,15 +1081,84 @@ class Simulation:
         return success, max_error_norm, min(relevant_scales)
 
 
-    def step_fixed(self, dt=None):
-        """Advances the simulation by one timestep 'dt' for fixed step solvers.
+    # timestepping ----------------------------------------------------------------
 
-        Selects between implicit and explicit solvers. Implicit solvers have 
-        an additional loop for solving the implicit update equation in each 
-        timestep.
+    def timestep_fixed_explicit(self, dt=None):
+        """Advances the simulation by one timestep 'dt' for explicit 
+        fixed step solvers.
 
-        If discrete events are detected, they are resolved immediately within 
-        the timestep.
+        If discrete events are detected, they are resolved immediately 
+        within the timestep.
+
+        Parameters
+        ----------
+        dt : float
+            timestep
+
+        Returns
+        -------
+        success : bool
+            indicator if the timestep was successful
+        max_error : float
+            maximum local truncation error from integration
+        scale : float
+            rescale factor for timestep
+        total_evals : int
+            total number of system evaluations
+        total_solver_its : int
+            total number of implicit solver iterations
+        """
+
+        #default global timestep as local timestep
+        if dt is None: 
+            dt = self.dt
+
+        #buffer internal states for solvers and event system
+        self._buffer(self.time, dt)
+
+        #total function evaluations 
+        total_evals = 0
+
+        #iterate explicit solver stages with evaluation time (generator)
+        for time_stage in self.engine.stages(self.time, dt):
+
+            #evaluate system equation by fixed-point iteration
+            total_evals += self._update(time_stage) 
+
+            #timestep for dynamical blocks (with internal states)
+            _1, error_norm, _3 = self._step(time_stage, dt)
+
+        #system time after timestep
+        time_dt = self.time + dt
+
+        #evaluate system equation before sampling and event check (+dt)
+        total_evals += self._update(time_dt) 
+
+        #handle events chronologically after timestep (+dt)
+        for event, _, ratio in self._detected_events(time_dt):
+
+            #fixed timestep -> resolve event directly
+            event.resolve(self.time + ratio * dt)  
+
+            #after resolve, evaluate system equation again -> propagate event
+            total_evals += self._update(time_dt)      
+
+        #sample data after successful timestep (+dt)
+        self._sample(time_dt)
+ 
+        #increment global time and continue simulation
+        self.time = time_dt 
+
+        #max local truncation error, timestep rescale, successful step
+        return True, error_norm, 1.0, total_evals, 0
+
+
+    def timestep_fixed_implicit(self, dt=None): 
+        """Advances the simulation by one timestep 'dt' for implicit 
+        fixed step solvers.
+
+        If discrete events are detected, they are resolved immediately 
+        within the timestep.
 
         Parameters
         ----------
@@ -1119,32 +1193,23 @@ class Simulation:
         total_evals, total_solver_its = 0, 0
 
         #iterate explicit solver stages with evaluation time (generator)
-        for time in self.engine.stages(self.time, dt):
+        for time_stage in self.engine.stages(self.time, dt):
 
-            #explicit solver stepping loop
-            if self.engine.is_explicit:
+            #solve implicit update equation and get iteration count
[…]
