### 2025-04-27 8e7b219
assemble lists of active system components before timestep instead of skipping inactive components in each loop, this is more efficient
