### 2025-04-27 f19e137
cleanup and refactoring helpers in Simulation
