### 2025-04-28 cf4d8a3
ensure backward compatibility for Simulation.step
