### 2025-04-29 0ce401e
added utils/graph.py with methods for analyzing a DAG in the pathsim context
