### 2025-04-29 f22f326
full refactor of simulation loop, added methods to compute algebraic depth and order blocks for execution
