+++ src/pathsim/simulation.py
@@ -18,6 +18,8 @@ import json
 import datetime
 import logging
 
+from collections import defaultdict
+
 from pathsim import __version__
 
 from ._constants import (
@@ -121,8 +123,6 @@ class Simulation:
         solver class for numerical integration from pathsim.solvers
     tolerance_fpi : float
         absolute tolerance for convergence of fixed-point iterations
-    iterations_min : int
-        minimum number of fixed-point iterations for system function evaluation
     iterations_max : int
         maximum allowed number of fixed-point iterations for system function evaluation
     log : bool, string
@@ -134,12 +134,6 @@ class Simulation:
     ----------
     time : float
         global simulation time
-    _active_blocks : list[Block]
-        list of active blocks to be included in simulation loop
-    _active_connections : list[Connection]
-        list of active connections to be included in simulation loop
-    _active_events : list[Event]
-        list of active events to be included in simulation loop
     path_length : int
         estimated length of the longest algebraic path
     engine : Solver
@@ -158,7 +152,6 @@ class Simulation:
         dt_max=SIM_TIMESTEP_MAX, 
         Solver=SSPRK22, 
         tolerance_fpi=SIM_TOLERANCE_FPI, 
-        iterations_min=SIM_ITERATIONS_MIN, 
         iterations_max=SIM_ITERATIONS_MAX, 
         log=LOG_ENABLE,
         **solver_kwargs
@@ -168,11 +161,6 @@ class Simulation:
         self.blocks      = []
         self.connections = []
         self.events      = []
-
-        #active system components (determined dynamically)
-        self._active_blocks      = []
-        self._active_connections = []
-        self._active_events      = []
 
         #simulation timestep and bounds
         self.dt     = dt
@@ -192,7 +180,6 @@ class Simulation:
         self.solver_kwargs = solver_kwargs
 
         #iterations for fixed-point loop
-        self.iterations_min = iterations_min
         self.iterations_max = iterations_max
 
         #enable logging flag
@@ -234,8 +221,8 @@ class Simulation:
         #set numerical integration solver
         self._set_solver()
 
-        #find length of longest algebraic path
-        self._algebraic_path_length()
+        #assemble the system for simulation
+        self._assemble()
 
 
     def __str__(self):
@@ -334,7 +321,7 @@ class Simulation:
         kwargs : dict
             kwargs for the plot method
         """
-        for block in self._active_blocks:
+        for block in self.blocks:
             block.plot(*args, **kwargs)
 
 
@@ -533,17 +520,9 @@ class Simulation:
         #add block to global blocklist
         self.blocks.append(block)
 
-        #add block to active list if active
-        if block:
-            self._active_blocks.append(block)
-
         #add events of block to global event list
         for event in block.get_events():
             self.add_event(event)
-
-        #recompute algebraic path length
-        if recompute_path:
-            self._algebraic_path_length()
 
 
     def add_connection(self, connection, recompute_path=True):
@@ -578,14 +557,6 @@ class Simulation:
         #add connection to global connection list
         self.connections.append(connection)
 
-        #add connection to active connection list if active
-        if connection:
-            self._active_connections.append(connection)
-
-        #recompute algebraic path length
-        if recompute_path:
-            self._algebraic_path_length()
-
 
     def add_event(self, event):
         """Checks and adds a new event to the simulation.
@@ -606,50 +577,78 @@ class Simulation:
         #add event to global event list
         self.events.append(event)
 
-        #add event to active event list if active
-        if event:
-            self._active_events.append(event)
+
+    # system assembly--------------------------------------------------------------
+
+
+    def _assemble(self):
+        self._assemble_components_alg_depth()
+        self._assemble_blocks_dyn()
+
+
[…]
