### 2025-04-29 0995d26
adapted all blocks with new update_err method, now convergence control for blocks algebraic components is only required when they are part of algebraic loops
