### 2025-04-29 3ae9aa1
added wrapper block
added Wrapper class to wrap any external function into pathsim, linked with a Shedule event to control the sampling  time
