### 2025-04-29 e811172
added Unittest for Wrapper block
