### 2025-04-29 cc4e532
Update __init__.py
Replace wrapper import for entropy minimization
