+++ src/pathsim/blocks/__init__.py
@@ -5,6 +5,7 @@ from .amplifier import *
 from .function import *
 from .spectrum import *
 from .sources import *
+from .wrapper import *
 from .switch import *
 from .adder import *
 from .scope import *
@@ -13,5 +14,3 @@ from .ctrl import *
 from .lti import *
 from .ode import *
 from .rng import *
-from .wrapper import *
-
