### 2025-04-30 bca443f
updated docstrings in _block
