### 2025-04-30 a9a0ad9
adapted Subsystem to mirror DAG magic from Simulation
