+++ src/pathsim/subsystem.py
@@ -14,11 +14,19 @@
 
 import numpy as np
 
+from collections import defaultdict
+
 from .connection import Connection
 
 from .blocks._block import Block
 
-from .utils.utils import path_length_dfs
+from .utils.graph import (
+    downstream_connection_map,
+    upstream_connection_map, 
+    upstream_path_length_dfs, 
+    path_length_dfs
+    )
+
 from .utils.register import Register
 from .utils.portreference import PortReference
 
@@ -39,6 +47,9 @@ class Interface(Block):
     - Handles data transfer to and from the internal subsystem blocks 
       to and from the inputs and outputs of the subsystem.
     """
+
+    def __len__(self):
+        return 0
     
     def set_output(self, port, value): 
         self.outputs[port] = value
@@ -158,12 +169,17 @@ class Subsystem(Block):
 
         Basically the same as in the 'Simulation' class.
         """
-        max_path_length = 0
-        for block in [self.interface, *self.blocks]:
-            path_length = path_length_dfs(self.connections, block)
-            if path_length > max_path_length:
-                max_path_length = path_length
-        return max_path_length
+
+        #build downstream connection map
+        dst_con_map = downstream_connection_map(self.connections)
+
+        #compute algebraic path from interface back to itself (internal path length)
+        length = path_length_dfs(dst_con_map, self.interface, self.interface)
+
+        #hit internal algebraic loop (inf)
+        if length is None: return None
+
+        return int(length > 0)
 
 
     def __call__(self):
@@ -218,6 +234,94 @@ class Subsystem(Block):
                 if conn_1.overwrites(conn_2):
                     _msg = f"{conn_1} overwrites {conn_2}"
                     raise ValueError(_msg)
+
+
+    # subsystem assembly ----------------------------------------------------------
+
+    def assemble(self):
+        """Assemble internals of subsystem for simulation.
+
+        Assemble internal blocks (recursively assemble nested subsystems)
+        and the internal DAG and dynamic blocks for timestepping
+        """
+
+        #assemble blocks internally
+        for block in self.blocks:
+            block.assemble()
+
+        #assemble system for simulation
+        self._assemble_components_alg_depth()
+        self._assemble_blocks_dyn()
+
+
+    def _assemble_components_alg_depth(self):
+        """Assemble lists of internal system components, ordered by 
+        their algebraic depth in the DAG.
+
+        Perform upstream depth first search on the DAG to assign 
+        each block their number of consecutive algebraic dependencies 
+        (algebraic depth). This is also used to assign the outgoing 
+        connections similarly.
+        """
+
+        #initial max algebraic depth (longest algebraic path)
+        max_depth = 0
+
+        #assembly of blocks by their algebraic depth
+        self._blocks_alg_depth = defaultdict(list)
+        self._blocks_alg_loop = []
+        
+        #assembly of connections by their algebraic depth
+        self._connections_alg_depth = defaultdict(list)
+        self._connections_alg_loop = []
+
+        #assembly of initerface connections
+        self._connections_interface = []
+
+        #construct mapping for connections between blocks
+        ust_con_map = upstream_connection_map(self.connections)
+
+        #construct mapping from blocks to outgoing connections
+        block_con_map = defaultdict(list)
+        for con in self.connections:
+            block_con_map[con.source.block].append(con)
+        
+        #iterate blocks to calculate their algebraic depths
+        for blk in self.blocks:
+            depth = upstream_path_length_dfs(ust_con_map, blk) 
+
+            #None -> alg. loop upstream taints downstream components
+            if depth is None:
+                self._blocks_alg_loop.append(blk)
+
+                #find outgoing connections and assign them aswell
+                for con in block_con_map[blk]:
[…]
