### 2025-04-30 6bb53ce
fixed alg loop update for Subsystem
