### 2025-04-30 850ad7f
added loop flag to assembly in Simulation
