### 2025-04-30 4433599
refactor of graphing, introduced utils/graph.py and Graph class that handles all graph operations, cleanup of Simulation and Subsystem
