+++ src/pathsim/simulation.py
@@ -32,8 +32,7 @@ from ._constants import (
     LOG_ENABLE
     )
 
-from .utils.graph import upstream_connection_map, upstream_path_length_dfs
-
+from .utils.graph import Graph
 from .utils.analysis import Timer
 from .utils.portreference import PortReference
 from .utils.progresstracker import ProgressTracker
@@ -172,6 +171,9 @@ class Simulation:
         #numerical integrator instance -> initialized later
         self.engine = None
 
+        #internal system graph -> initialized later
+        self.graph = None
+
         #error tolerance for fixed point loop and implicit solver
         self.tolerance_fpi = tolerance_fpi
 
@@ -186,9 +188,6 @@ class Simulation:
 
         #initial simulation time
         self.time = 0.0
-
-        #length of the longest algebraic path
-        self.path_length = 1
 
         #initialize logging for logging mode
         self._initialize_logger()
@@ -217,11 +216,11 @@ class Simulation:
         #check if blocks from connections are in simulation
         self._check_blocks_are_managed()
 
-        #set numerical integration solver
+        #set numerical integration solver (block local)
         self._set_solver()
 
-        #assemble the system for simulation
-        self._assemble()
+        #assemble the system graph for simulation
+        self._assemble_graph()
 
 
     def __str__(self):
@@ -578,87 +577,17 @@ class Simulation:
 
     # system assembly -------------------------------------------------------------
 
-    def _assemble(self):
+    def _assemble_graph(self):
 
         #assemble blocks internally
         for block in self.blocks:
             block.assemble()
 
-        #assemble system for simulation
-        self._assemble_components_alg_depth()
-        self._assemble_blocks_dyn()
+        #assemble graph for simulation
+        self.graph = Graph(self.blocks, self.connections)
+        self._logger_info(f"GRAPH (depth: {len(self.graph)}, loops: {self.graph.has_loops})")
 
 
-    def _assemble_components_alg_depth(self):
-        """Assemble lists of system components, ordered by their 
-        algebraic depth in the DAG.
-
-        Perform upstream depth first search on the DAG to assign 
-        each block their number of consecutive algebraic dependencies 
-        (algebraic depth). This is also used to assign the outgoing 
-        connections similarly.
-        """
-
-        #initial max algebraic depth (longest algebraic path)
-        max_depth = 0
-
-        #loop detection flag
-        loops = False
-
-        #assembly of blocks by their algebraic depth
-        self._blocks_alg_depth = defaultdict(list)
-        self._blocks_alg_loop = []
-        
-        #assembly of connections by their algebraic depth
-        self._connections_alg_depth = defaultdict(list)
-        self._connections_alg_loop = []
-
-        #construct mapping for connections between blocks
-        ust_con_map = upstream_connection_map(self.connections)
-
-        #construct mapping from blocks to outgoing connections
-        block_con_map = defaultdict(list)
-        for con in self.connections:
-            block_con_map[con.source.block].append(con)
-        
-        #iterate blocks to calculate their algebraic depths
-        for blk in self.blocks:
-            depth = upstream_path_length_dfs(ust_con_map, blk) 
-
-            #None -> alg. loop upstream taints downstream components
-            if depth is None:
-                self._blocks_alg_loop.append(blk)
-
-                #find outgoing connections and assign them aswell
-                for con in block_con_map[blk]:
-                    self._connections_alg_loop.append(con)
-
-                #loop detection flag
-                loops = True
-
-            else:
-                self._blocks_alg_depth[depth].append(blk)
-
-                #find outgoing connections and assign them aswell
-                for con in block_con_map[blk]:
-                    self._connections_alg_depth[depth].append(con)
-
-                #update max depth
[…]
