+++ src/pathsim/utils/graph.py
@@ -9,7 +9,7 @@
 
 # IMPORTS ==============================================================================
 
-from collections import defaultdict
+from collections import defaultdict, deque, namedtuple
 from functools import cache
 
 
@@ -70,7 +70,6 @@ def upstream_block_block_map(connections):
     -------
     connection_map : defaultdict[Block, set[Block]]
         directed upstream graph of connections (target -> source)
-
     """
     connection_map = defaultdict(set)
     for con in connections:
@@ -81,8 +80,7 @@ def upstream_block_block_map(connections):
 
 
 def _block_meta(block):
-    """
-    Return a triple: (alg_len, is_hidden_loop, is_algebraic)
+    """Return a triple: (alg_len, is_hidden_loop, is_algebraic)
 
     * `len(block) is None` -> (0,  True,  False)
     * `len(block) == 0`    -> (0,  False, False)
@@ -177,8 +175,7 @@ def _dfs_generic(graph_map, node, propagate_inf, start, stack=frozenset()):
 
 
 def upstream_path_length_dfs(connection_map, starting_block):
-    """
-    Longest algebraic path length ending at `starting_block`, walking
+    """Longest algebraic path length ending at `starting_block`, walking
     **upstream** (targets -> sources).
 
     * Stops – and finishes that branch – as soon as a *non-algebraic* block
@@ -199,9 +196,8 @@ def upstream_path_length_dfs(connection_map, starting_block):
 
 
 def downstream_path_length_dfs(connection_map, starting_block):
-    """
-    Longest algebraic path length starting from `starting_block`, walking
-    **downstream** (source -> targets).
+    """Longest algebraic path length starting from `starting_block`, 
+    walking **downstream** (source -> targets).
 
     * Traversal *immediately* stops on a non-algebraic block, a hidden-loop
       block, or a loop farther downstream; those branches contribute 0.
@@ -220,11 +216,10 @@ def downstream_path_length_dfs(connection_map, starting_block):
     )
 
 
-def path_length_dfs(connection_map, start_block, end_block):
-    """
-    Return the length of the *longest purely-algebraic* directed path from
-    `src_block` to `dst_block` following the `connection_map` orientation
-    (source -> targets).
+def distance_path_length_dfs(connection_map, start_block, end_block):
+    """Return the length of the *longest purely-algebraic* directed path 
+    from `src_block` to `dst_block` following the `connection_map` 
+    orientation (source -> targets).
 
     Parameters
     ----------
@@ -293,12 +288,17 @@ def path_length_dfs(connection_map, start_block, end_block):
 # GRAPH CLASS ==========================================================================
 
 class Graph:
-
     """Representation of a directed graph, defined by blocks (nodes) 
     and edges (connections).
 
     Manages graph analysis methods.
 
+    Parameter
+    ---------
+    blocks : list[blocks] | None
+        blocks / nodes of the graph
+    connections : list[Connection] | None
+        connections / edges of the graph
     """
 
     def __init__(self, blocks=None, connections=None):
@@ -311,10 +311,10 @@ class Graph:
 
         #initialize graph orderings
         self._blocks_dag = defaultdict(list)
-        self._blocks_loop = []
+        self._blocks_loop_dag = defaultdict(list)
         
         self._connections_dag = defaultdict(list)
-        self._connections_loop = []
+        self._connections_loop_dag = defaultdict(list)
 
         #construct mappings for connections between blocks
         self.upst_blk_blk_map = upstream_block_block_map(self.connections)
@@ -330,7 +330,80 @@ class Graph:
 
 
     def __len__(self):
-        return self._alg_depth
+        return len(self.blocks)
+
+    def _build_loop_depths(self, blocks_loop):
[…]
