### 2025-04-30 223e89b
refactor, adding error_control=False as default arg to Block.update to make error propagation optional, only relevant for algebraic loop solver
