+++ examples/example_solver_hotswap.py
@@ -49,9 +49,9 @@ Sim = Simulation(
     connections, 
     dt=0.5, 
     log=True, 
-    tolerance_lte_abs=1e-6, 
+    tolerance_lte_abs=1e-5, 
     tolerance_lte_rel=1e-3,
-    tolerance_fpi=1e-9
+    tolerance_fpi=1e-8
     )
 
 
