+++ src/pathsim/blocks/ctrl.py
@@ -131,28 +131,15 @@ class PID(Block):
         else: self.engine = Solver.cast(self.engine, **solver_args)
 
 
-    def update(self, t):
-        """update system equation fixed point loop,
-        without convergence control
+    def update(self, t, error_control=False):
+        """update system equation fixed point loop, with convergence control
     
         Parameters
         ----------
         t : float
             evaluation time
-        """
-        x, u = self.engine.get(), self.inputs[0]
-        y = self.op_alg(x, u, t)
-        self.outputs.update_from_array(y)    
-
-
-    def update_err(self, t):
-        """update system equation fixed point loop,
-        with convergence control
-    
-        Parameters
-        ----------
-        t : float
-            evaluation time
+        error_control : bool
+            activate error control 
 
         Returns
         -------
@@ -161,7 +148,10 @@ class PID(Block):
         """
         x, u = self.engine.get(), self.inputs[0]
         y = self.op_alg(x, u, t)
-        return self.outputs.update_from_array_max_err(y)
+        if error_control:
+            return self.outputs.update_from_array_max_err(y)
+        self.outputs.update_from_array(y) 
+        return 0.0
 
 
     def solve(self, t, dt):
