+++ src/pathsim/blocks/multiplier.py
@@ -52,25 +52,15 @@ class Multiplier(Block):
             )
 
 
-    def update(self, t):
+    def update(self, t, error_control=False):
         """update system equation
 
         Parameters
         ----------
         t : float
             evaluation time
-        """
-        u = self.inputs.to_array()
-        self.outputs.update_from_array(self.op_alg(u))
-
-
-    def update_err(self, t):
-        """update system equation in with error control for algebraic loop solver
-
-        Parameters
-        ----------
-        t : float
-            evaluation time
+        error_control : bool
+            activate error control 
 
         Returns
         -------
@@ -78,4 +68,7 @@ class Multiplier(Block):
             absolute error to previous iteration for convergence control
         """
         u = self.inputs.to_array()
-        return self.outputs.update_from_array_max_err(self.op_alg(u))
+        if error_control:
+            return self.outputs.update_from_array_max_err(self.op_alg(u))
+        self.outputs.update_from_array(self.op_alg(u))
+        return 0.0
