### 2025-04-30 1c6bb6c
Update test_wrapper.py
corrected test for internal event
