### 2025-05-01 607e874
refactor, back to original signature of the Block.update method, performance impact of internal checks was similar to always performing error estimates for algebraic loops
