### 2025-05-01 cca33f5
cleanup and docstrings
