+++ src/pathsim/simulation.py
@@ -112,32 +112,49 @@ class Simulation:
     events : list[Event]
         list of event trackers (zero crossing detection, schedule, etc.)
     dt : float
-        transient simulation timestep in time units
+        transient simulation timestep in time units, 
+        default see ´SIM_TIMESTEP´ in ´_constants.py´
     dt_min : float
-        lower bound for transient simulation timestep, default '0.0'
+        lower bound for transient simulation timestep, 
+        default see ´SIM_TIMESTEP_MIN´ in ´_constants.py´
     dt_max : float
-        upper bound for transient simulation timestep, default 'None'
+        upper bound for transient simulation timestep, 
+        default see ´SIM_TIMESTEP_MAX´ in ´_constants.py´
     Solver : Solver 
-        solver class for numerical integration from pathsim.solvers
+        ODE solver class for numerical integration from ´pathsim.solvers´,
+        default is ´pathsim.solvers.ssprk22.SSPRK22´ (2nd order expl. Runge Kutta)
     tolerance_fpi : float
-        absolute tolerance for convergence of fixed-point iterations
+        absolute tolerance for convergence of algebraic loops 
+        and internal optimizers of implicit ODE solvers, 
+        default see ´SIM_TOLERANCE_FPI´ in ´_constants.py´
     iterations_max : int
-        maximum allowed number of fixed-point iterations for system function evaluation
-    log : bool, string
-        flag to enable logging (alternatively a path to a log file can be specified)
+        maximum allowed number of iterations for implicit ODE 
+        solver optimizers and algebraic loop solver, 
+        default see ´SIM_ITERATIONS_MAX´ in ´_constants.py´
+    log : bool | string
+        flag to enable logging, default see ´LOG_ENABLE´ in ´_constants.py´
+        (alternatively a path to a log file can be specified)
     solver_kwargs : dict
-        additional parameters for numerical solvers such as abs and rel tolerance
+        additional parameters for numerical solvers such as absolute 
+        (´tolerance_lte_abs´) and relative (´tolerance_lte_rel´) tolerance, 
+        defaults are defined in ´_constants.py´
 
     Attributes
     ----------
     time : float
-        global simulation time
-    path_length : int
-        estimated length of the longest algebraic path
+        global simulation time, starting at ´0.0´
+    graph : Graph
+        internal graph representation for fast system funcion 
+        evluations using DAG with algebraic depths
     engine : Solver
-        global integrator instance
+        global integrator (ODE solver) instance serving as a dummy to 
+        get attributes and access to intermediate evaluation stages
     logger : logging.Logger
         global simulation logger
+    _needs_buffering : bool
+        flag for buffering system state
+    _blocks_dyn : list[Block]
+        list of blocks with internal ´Solver´ instances (stateful) 
     """
 
     def __init__(
@@ -168,8 +185,8 @@ class Simulation:
         #numerical integrator to be used (class definition)
         self.Solver = Solver
 
-        #numerical integrator instance -> initialized later
-        self.engine = None
+        #numerical integrator instance
+        self.engine = Solver()
 
         #internal system graph -> initialized later
         self.graph = None
@@ -189,24 +206,24 @@ class Simulation:
         #initial simulation time
         self.time = 0.0
 
+        #flag for state buffering (transient)
+        self._needs_buffering = True
+
+        #collection of blocks with internal ODE solvers
+        self._blocks_dyn = []
+
         #initialize logging for logging mode
         self._initialize_logger()
 
         #prepare and add blocks (including internal events)
         if blocks is not None:
             for block in blocks:
-                self.add_block(
-                    block, 
-                    recompute_path=False
-                    )
+                self.add_block(block)
 
         #check and add connections
         if connections is not None:
             for connection in connections:
[…]
