### 2025-05-01 8d59ac6
cleanup and docstrings for Graph
