+++ src/pathsim/utils/graph.py
@@ -291,14 +291,39 @@ class Graph:
     """Representation of a directed graph, defined by blocks (nodes) 
     and edges (connections).
 
-    Manages graph analysis methods.
+    Manages graph analysis methods, detect algebraic loops, sort blocks and 
+    connections into algebraic levels by depth in directed acyclic graph (DAG). 
+    Does the same for algebraic loop blocks and connections (DAG for open loops).
 
-    Parameter
-    ---------
+    Parameters
+    ----------
     blocks : list[blocks] | None
         blocks / nodes of the graph
     connections : list[Connection] | None
         connections / edges of the graph
+
+    Attributes
+    ----------
+    has_loops : bool
+        flag to indicate if graph has algebraic loops
+    _alg_depth : int
+        algebraic depth of DAG (number of alg. levels)
+    _loop_depth : int
+        algebraic depth of broken loop DAG (number of alg. levels in loops)
+    _blocks_dag : defaultdict[int, list[Block]]
+        algebraic levels of blocks in internal DAG
+    _blocks_loop_dag : defaultdict[int, list[Block]]
+        algebraic levels of blocks in broken loops DAG
+    _connections_dag : defaultdict[int, list[Connection]]
+        algebraic levels of connections in internal DAG
+    _connections_loop_dag : defaultdict[int, list[Connection]]
+        algebraic levels of connections in broken loops DAG
+    _upst_blk_blk_map : defaultdict[Block, set[Block]]
+        map for upstream connections between blocks
+    _dnst_blk_blk_map : defaultdict[Block, set[Block]]
+        map for downstream connections between blocks
+    _outg_blk_con_map : defaultdict[Block, list[Connection]]
+        map for outgoing connections of blocks
     """
 
     def __init__(self, blocks=None, connections=None):
@@ -309,6 +334,10 @@ class Graph:
         #loop flag
         self.has_loops = False
 
+        #depths
+        self._alg_depth = 0
+        self._loop_depth = 0
+
         #initialize graph orderings
         self._blocks_dag = defaultdict(list)
         self._blocks_loop_dag = defaultdict(list)
@@ -317,9 +346,9 @@ class Graph:
         self._connections_loop_dag = defaultdict(list)
 
         #construct mappings for connections between blocks
-        self.upst_blk_blk_map = upstream_block_block_map(self.connections)
-        self.dnst_blk_blk_map = downstream_block_block_map(self.connections)
-        self.outg_blk_con_map = outgoing_block_connection_map(self.connections)
+        self._upst_blk_blk_map = upstream_block_block_map(self.connections)
+        self._dnst_blk_blk_map = downstream_block_block_map(self.connections)
+        self._outg_blk_con_map = outgoing_block_connection_map(self.connections)
 
         #assemble dag and loops
         self._assemble()
@@ -331,6 +360,11 @@ class Graph:
 
     def __len__(self):
         return len(self.blocks)
+
+
+    def depth(self):
+        return self._alg_depth, self._loop_depth
+
 
     def _build_loop_depths(self, blocks_loop):
         """Populate the defaultdicts that order **only the blocks/connections
@@ -360,15 +394,15 @@ class Graph:
         blocks in Gauss–Seidel order (shallow -> deep) if desired.
         """
 
-        # safety: nothing to do
-        if not blocks_loop:  return
+        #safety, nothing to do
+        if not blocks_loop: return
 
         loop_set = set(blocks_loop)
         depth_of  = {}
 
         #collect entry nodes
         for blk in loop_set:
-            preds = self.upst_blk_blk_map[blk]
+            preds = self._upst_blk_blk_map[blk]
             if not preds or any(p not in loop_set for p in preds):
[…]
