### 2025-05-01 a1ddd7f
cleanup of constants
