+++ src/pathsim/_constants.py
@@ -13,12 +13,11 @@ TOLERANCE = 1e-16
 
 # simulation default constants ---------------------------------------------------
 
-SIM_TIMESTEP = 0.01         # simulator timestep (initial) 
-SIM_TIMESTEP_MIN = 1e-16    # minimum simulator timestep
-SIM_TIMESTEP_MAX =  None    # maximum simulator timestep
-SIM_TOLERANCE_FPI = 1e-12   # tolerance for fixed point loop convergence
-SIM_ITERATIONS_MIN = None   # minimum number of fixed point loop iterations -> determined automatically
-SIM_ITERATIONS_MAX = 200    # maximum number of fixed point loop iterations
+SIM_TIMESTEP = 0.01         # transient simulation timestep (initial) 
+SIM_TIMESTEP_MIN = 1e-16    # min allowed transient timestep
+SIM_TIMESTEP_MAX =  None    # max allowed transient timestep
+SIM_TOLERANCE_FPI = 1e-12   # tolerance for optimizer / alg. loop solver
+SIM_ITERATIONS_MAX = 200    # max number of optimizer / alg. loop solver iterations
 
 
 # solver default constants -------------------------------------------------------
@@ -26,15 +25,15 @@ SIM_ITERATIONS_MAX = 200    # maximum number of fixed point loop iterations
 SOL_TOLERANCE_LTE_ABS = 1e-8   # absolute local truncation error (adaptive solvers)
 SOL_TOLERANCE_LTE_REL = 1e-5   # relative local truncation error (adaptive solvers)
 SOL_TOLERANCE_FPI = 1e-9       # tolerance for optimizer convergence (implicit solvers)
-SOL_ITERATIONS_MAX = 200       # maximum number of optimizer iterations (implicit solvers)
-SOL_SCALE_MIN = 0.1            # minimum timestep rescale factor (adaptive solvers)
-SOL_SCALE_MAX = 10             # maximum timestep rescale factor (adaptive solvers)
+SOL_ITERATIONS_MAX = 200       # max number of optimizer iterations (for standalone implicit solvers)
+SOL_SCALE_MIN = 0.1            # min allowed timestep rescale factor (adaptive solvers)
+SOL_SCALE_MAX = 10             # max allowed timestep rescale factor (adaptive solvers)
 SOL_BETA = 0.9                 # savety for timestep control (adaptive solvers)
 
 
 # event default constants --------------------------------------------------------
 
-EVT_TOLERANCE = 1e-4   # tolerance for event detection
+EVT_TOLERANCE = 1e-4   # tolerance for event detection (zero-crossing, condition)
 
 
 # logging default constants ------------------------------------------------------
