### 2025-05-01 86949ff
added __float__ to Timer
