+++ src/pathsim/_constants.py
@@ -11,6 +11,7 @@
 
 TOLERANCE = 1e-16
 
+
 # simulation default constants ---------------------------------------------------
 
 SIM_TIMESTEP = 0.01         # transient simulation timestep (initial) 
