### 2025-05-01 8d72b66
improved logging messages for Simulation
