### 2025-05-01 193b44c
added example for cascade control loop with noise
