### 2025-05-01 21aea07
simplified __contains__ for Simulation and Subsystem, fixed typos in docstrings
