### 2025-05-01 #8097 @RemDelaporteMathurin
@danielskatz although we haven't published our `pathsim` based results yet, we are using it for several chemical applications (fuel cycle of fusion power plants)
