### 2025-05-01 #8097 @RemDelaporteMathurin
@milanofthe I'd be happy to share the ARC fuel cycle demo I have to include in this paper if that helps
