+++ tests/pathsim/test_subsystem.py
@@ -106,37 +106,6 @@ class TestSubsystem(unittest.TestCase):
             S = Subsystem(blocks=[B1, B2, B3, I1], connections=[C1, C2, C3])
 
 
-    def test_len(self):
-
-        #test the len method for internal signal path estimation
-
-        I1 = Interface()
-        S = Subsystem(blocks=[I1],)
-        self.assertEqual(len(S), 1)
-
-        B1 = Block()
-        I1 = Interface()
-        C1 = Connection(I1, B1)
-        S = Subsystem(blocks=[I1, B1], connections=[C1])
-        self.assertEqual(len(S), 2)
-
-        B1, B2 = Block(), Block()
-        I1 = Interface()
-        C1 = Connection(I1, B1)
-        C2 = Connection(B1, B2)
-        S = Subsystem(blocks=[I1, B1, B2], connections=[C1, C2])
-        self.assertEqual(len(S), 3)
-
-        B1, B2, B3 = Block(), Block(), Block()
-        I1 = Interface()
-        C1 = Connection(I1, B1, B2, B3)
-        C2 = Connection(B1, B2[1])
-        C3 = Connection(B2, B1[1], B3[1])
-        C4 = Connection(B3, I1)
-        S = Subsystem(blocks=[I1, B1, B2, B3], connections=[C1, C2, C3, C4])
-        self.assertEqual(len(S), 4)
-
-
     def test_set(self): 
 
         B1 = Block()
@@ -184,39 +153,11 @@ class TestSubsystem(unittest.TestCase):
         self.assertEqual(err, 0.0)
 
 
-    def test_nesting(self):
+    # to be implemented ----------------------------------------------------------------
 
-        #nesting depth 0
-        B1 = Block()
-        I1 = Interface()
-        C1 = Connection(I1, B1)
-        C2 = Connection(B1, I1)
-
-        S1 = Subsystem(blocks=[I1, B1], connections=[C1, C2])
-
-        self.assertEqual(len(S1), 2)
-
-        #nesting depth 1
-        B2 = Block()
-        I2 = Interface()
-        C3 = Connection(I2, S1)
-        C4 = Connection(S1, B2)
-        C5 = Connection(B2, I2)
-
-        S2 = Subsystem(blocks=[I2, B2, S1], connections=[C3, C4, C5])
-
-        self.assertEqual(len(S2), 4)
-
-        #nesting depth 2
-        B3 = Block()
-        I3 = Interface()
-        C6 = Connection(I3, S2)
-        C7 = Connection(S2, B3)
-        C8 = Connection(B3, I3)
-
-        S3 = Subsystem(blocks=[I3, B3, S2], connections=[C6, C7, C8])
-
-        self.assertEqual(len(S3), 6)
+    def test_nesting(self): pass
+    def test_len(self): pass
+    def test_graph(self): pass
 
 
 # RUN TESTS LOCALLY ====================================================================
