### 2025-05-01 e38ca74
added tests for Graph in tests/pathsim/utils/test_graph.py
