+++ tests/pathsim/utils/test_graph.py
@@ -1,0 +1,219 @@
+########################################################################################
+##
+##                                   TESTS FOR 
+##                                'utils/graph.py'
+##
+##                               Milan Rother 2025
+##
+########################################################################################
+
+# IMPORTS ==============================================================================
+
+import unittest
+import numpy as np
+
+from pathsim.utils.graph import Graph
+from pathsim.blocks import Block, Integrator, Amplifier, Adder
+from pathsim.connection import Connection
+
+
+# TESTS ================================================================================
+
+class TestGraph(unittest.TestCase):
+    """tests for the Graph class in pathsim.utils.graph"""
+
+    def setUp(self):
+        """Set up common resources for test methods."""
+        # Algebraic Blocks (assume len=1)
+        self.amp1 = Amplifier(gain=2) # len=1
+        self.amp2 = Amplifier(gain=-1) # len=1
+        self.add1 = Adder() # len=1
+        self.add2 = Adder("+-") # Subtractor, len=1
+
+        # Dynamic Blocks (assume len=0)
+        self.int1 = Integrator(initial_value=0.0) # len=0
+        self.int2 = Integrator(initial_value=1.0) # len=0
+
+        # --- Connections for Acyclic Mixed Graph ---
+        # Longest algebraic path: amp1(1) -> add1(1) -> amp2(1) = length 3, ends at depth 3
+        # Graph depth = max_node_depth + 1 = 3 + 1 = 4
+        self.c_amp1_add1 = Connection(self.amp1, self.add1[0])
+        self.c_int1_add1 = Connection(self.int1, self.add1[1])
+        self.c_add1_amp2 = Connection(self.add1, self.amp2)
+        self.c_amp2_int2 = Connection(self.amp2, self.int2)
+        self.nodes_acyclic = [self.amp1, self.add1, self.int1, self.amp2, self.int2]
+        self.conns_acyclic = [self.c_amp1_add1, self.c_int1_add1, self.c_add1_amp2, self.c_amp2_int2]
+
+        # --- Connections for Algebraic Loop Graph ---
+        # Loop: add2 -> amp1 -> add2
+        self.c_add2_amp1 = Connection(self.add2, self.amp1)
+        self.c_amp1_add2 = Connection(self.amp1, self.add2[1])
+        self.nodes_alg_loop = [self.add2, self.amp1]
+        self.conns_alg_loop = [self.c_add2_amp1, self.c_amp1_add2]
+
+        # --- Connections for Dynamic Loop Graph ---
+        # Loop: int1 -> int2 -> int1
+        self.c_int1_int2 = Connection(self.int1, self.int2)
+        self.c_int2_int1 = Connection(self.int2, self.int1)
+        self.nodes_dyn_loop = [self.int1, self.int2]
+        self.conns_dyn_loop = [self.c_int1_int2, self.c_int2_int1]
+
+    # --- Initialization Tests ---
+
+    def test_init_empty(self):
+        """Test initializing an empty graph."""
+        g = Graph([], [])
+        self.assertEqual(len(g), 0)
+        self.assertFalse(g.has_loops)
+        alg_depth, loop_depth = g.depth()
+        self.assertEqual(alg_depth, 0)
+        self.assertEqual(loop_depth, 0)
+
+    def test_init_with_mixed_nodes_edges_acyclic(self):
+        """Test initializing a graph with mixed nodes and acyclic connections."""
+        g = Graph(self.nodes_acyclic, self.conns_acyclic)
[…]
