### 2025-05-01 2eebe3b
updated tests for Graph
