### 2025-05-01 8c87b69
added PortReference.to(other) method for direct data transfer between blocks, adapted Connection.update to use this cleaner version
