### 2025-05-02 9954843
cleanup and adding __slots__ to Connection and PortReference
