### 2025-05-02 f0abe18
added comments in Wrapper class
added description of Wrapper class use
