### 2025-05-02 b84b7dc
added Block.size, Simulation.size and Subsystem.size to get number of blocks and internal dynamical states
