+++ examples/examples_odes/example_duffing.py
@@ -77,7 +77,7 @@ Sim = Simulation(
 
 if __name__ == "__main__":
 
-    Sim.run(20)
+    Sim.run(200)
 
     #plot the results directly from the scope
     Sc.plot()
