### 2025-05-02 ceb2325
fixed typo in Block.size
