### 2025-05-02 ec358a8
expanded docstrings of Function and Source block to explain they can be used as decorators
