### 2025-05-02 1915c69
added model with algebraic loop to tests/models/_models
