### 2025-05-02 #8158 @RemDelaporteMathurin
@milanofthe instead of adding the code block as a png, could you add it as an actual code block in markdown?
