### 2025-05-02 #8158 @milanofthe
@RemDelaporteMathurin yes, do you mean similar to the readme, splitting up visuals and code?
