### 2025-05-03 3705375
added tests for Simulation.size and Simulation.__contains__
