### 2025-05-03 300d5bd
added tests for Subsystem.__contains__, Subsystem.__len__, Subsystem.size
