+++ tests/pathsim/test_subsystem.py
@@ -15,7 +15,7 @@ import numpy as np
 from pathsim.subsystem import Subsystem, Interface
 
 #for testing
-from pathsim.blocks._block import Block
+from pathsim.blocks import Block
 from pathsim.connection import Connection
 
 
@@ -155,9 +155,86 @@ class TestSubsystem(unittest.TestCase):
 
     # to be implemented ----------------------------------------------------------------
 
-    def test_nesting(self): pass
-    def test_len(self): pass
+    def test_contains(self):
+
+        B1, B2, B3 = Block(), Block(), Block()
+        I1 = Interface()
+        C1 = Connection(I1, B1, B2, B3)
+        C2 = Connection(B1, I1)
+        S = Subsystem(
+            blocks=[B1, B2, B3, I1], 
+            connections=[C1]
+            )
+
+        self.assertTrue(B1 in S)
+        self.assertTrue(B2 in S)
+        self.assertTrue(B3 in S)
+
+        self.assertTrue(C1 in S)
+        self.assertFalse(C2 in S)
+
+
+    def test_size(self):   
+
+        #test 3 alg. blocks
+        I1 = Interface()
+        B1, B2, B3 = Block(), Block(), Block()
+        C1 = Connection(B1, B2)
+        C2 = Connection(B2, B3)
+        C3 = Connection(B3, B1)
+        S = Subsystem(
+            blocks=[I1, B1, B2, B3], 
+            connections=[C1, C2, C3]
+            )  
+
+        n, nx = S.size()
+        self.assertEqual(n, 3)
+        self.assertEqual(nx, 0)
+
+        #test 1 dyn, 1 alg block
+        from pathsim.blocks import Integrator
+
+        I1 = Interface()
+        B1, B2 = Block(), Integrator(3)
+        C1 = Connection(B1, B2)
+        S = Subsystem(
+            blocks=[I1, B1, B2], 
+            connections=[C1]
+            )  
+
+
+        n, nx = S.size()
+        self.assertEqual(n, 2)
+        self.assertEqual(nx, 0) # <- no internal engine yet
+
+        from pathsim.solvers import EUF
+        S.set_solver(EUF)
+
+        n, nx = S.size()
+        self.assertEqual(nx, 1)
+
+
+    def test_len(self): 
+
+        I1 = Interface()
+        B1 = Block()
+        C1 = Connection(I1, B1)
+        C2 = Connection(B1, I1)
+        S = Subsystem(
+            blocks=[I1, B1], 
+            connections=[C1, C2]
+            ) 
+
+        #should be 1
+        self.assertEqual(len(S), 0)
+
+
+
+
     def test_graph(self): pass
+    def test_nesting(self): pass
+
+
 
 
 # RUN TESTS LOCALLY ====================================================================
