### 2025-05-03 104a837
refactor of the Graph class and its helper methods, much cleaner, now uses one unified dfs algorithm for DAG construction
