+++ src/pathsim/utils/graph.py
@@ -79,75 +79,71 @@ def upstream_block_block_map(connections):
     return connection_map
 
 
-def _block_meta(block):
-    """Return a triple: (alg_len, is_hidden_loop, is_algebraic)
-
-    * `len(block) is None` -> (0,  True,  False)
-    * `len(block) == 0`    -> (0,  False, False)
-    * `len(block) > 0`     -> (len, False, True)
-    """
-    raw = len(block)
-    if raw is None:
-        return 0, True, False
-    return raw, False, raw > 0
-
-
-def _dfs_generic(graph_map, node, propagate_inf, start, stack=frozenset()):
-    """Internal DFS that implements all path-length semantics.
-
+def algebraic_depth_dfs(graph_map, node, node_set=None, propagate_inf=True, stack=frozenset()):
+    """Computes the longest algebraic path length using depth-first search.
+    
+    This function determines the longest algebraic path length from a given node,
+    traversing the graph in the direction specified by the graph_map. It can
+    operate in two modes controlled by the propagate_inf parameter:
+    
+    - In propagate mode (propagate_inf=True): Cycles and hidden loops cause the
+      entire result to be marked as infinite (None). Used for detecting algebraic
+      loops that affect the entire system.
+    - In non-propagate mode (propagate_inf=False): Cycles and hidden loops act as
+      termination points (return 0). Used for constructing a DAG from components
+      that would otherwise form a cycle.
+    
     Parameters
     ----------
     graph_map : dict[Block, set[Block]]
-        Neighbour map *in the traversal direction*:
-        * upstream call  -> {target : {sources}}
-        * downstream call -> {source : {targets}}
-
+        Adjacency map in the direction of traversal:
+        - For upstream traversal: {target : {sources}}
+        - For downstream traversal: {source : {targets}}
     node : Block
-        Current block in the recursion.
-
-    propagate_inf : bool
-        *True*  ->  an 'infinite' branch (algebraic loop or hidden-loop block)
-        propagates upstream to taint the **entire** result (`None`).
-        (*Used by upstream_path_length_dfs*.)
-
-        *False* ->  an infinite branch is clipped locally; the overall
-        result is only `None` if the **start** block itself lies in/with
-        an algebraic loop or is a hidden-loop block.
-        (*Used by downstream_path_length_dfs*.)
-
-    start : Block
-        Block for which the public API was invoked; needed to recognise
-        whether a detected cycle involves the start node.
-
+        Starting block for the traversal.
+    node_set : set[Block] | None, optional
+        Optional set of blocks to restrict the search to.
+        If None, all blocks in the graph_map are considered.
+    propagate_inf : bool, default=True
+        Controls how cycles and hidden loops are handled:
+        - True: Infinite paths propagate and taint the entire result (None)
+        - False: Infinite paths simply terminate that branch (0)
     stack : frozenset[Block], optional
-        Blocks on the current recursion path (immutable for hashability).
-        Do **not** supply manually; the first call leaves this argument
-        at its default.
-
+        Blocks on the current recursion path (used for cycle detection).
+        Internal use only - do not supply manually.
+    
     Returns
     -------
     int | None
-        * Positive integer – longest finite algebraic path length.
-        * 0 – no algebraic blocks reachable before a path-breaking
-          non-algebraic block.
-        * None – result is “infinite” under the rules described above.
+        - Positive integer: Length of longest finite algebraic path
+        - 0: No algebraic path exists (all paths broken by non-algebraic blocks)
+        - None: Infinite path length (due to algebraic loops or hidden loops)
     """
     @cache
     def dfs(cur, stk):
-        alg_len, hidden_loop, is_alg = _block_meta(cur)
 
-        #hidden loop block
-        if hidden_loop:
-            return None if (propagate_inf or cur is start) else 0
[…]
