### 2025-05-03 milanofthe/pathsim#27 @sea-bass
[JOSS Review] Add community guidelines
From the JOSS review guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Could you add a contributor guide either as a markdown in this repo or in the ReadTheDocs documentation? (or both!)

https://github.com/openjournals/joss-reviews/issues/8158
