### 2025-05-03 milanofthe/pathsim#28 @sea-bass
[JOSS Review] Add more information and doc example for subsystems
I was reading the documentation before I downloaded the software, and was curious whether PathSim had subsystem functionality.

Nowhere in the docs outside the API docs was this very clear, although there are great examples already available. I think the documentation should explicitly have a subsystem example as part of its featured examples, to make this clear to users.

As a secondary request, it might be beneficial to explicitly call out subsystems in the API reference for system definition at https://pathsim.readthedocs.io/en/latest/api.html#system-definition-and-simulation. Maybe some slight modification to the diagram and a little more text?

https://github.com/openjournals/joss-reviews/issues/8158
