### 2025-05-03 milanofthe/pathsim#29 @sea-bass
[JOSS Review] Paper Feedback
This issue contains all my feedback for the paper submission to JOSS. Overall, the paper is in fantastic shape and I only have minor comments.

- [x] "Matlab Simulink" is the incorrect styling; it should be MathWorks Simulink. Also the citation reads weird with the "n.d-a, n.d.-b". Appended to it. Consider cleaning up the source file for the bibliography to address this
  - Although looking at the citations at the end, I see that your intent was to cite both MATLAB and Simulink, in which case I might consider having a citation that looks more like "MATLAB and Simulink (The MathWorks, Inc.)"

- [x] Nitpick: in the first sentence it reads "control engineering, chemical process engineering, OR mixed engineering." The "or" should be an "and".
- [x] Typo (line 25): "seemlessly" -> "seamlessly"
- [x] Typo (line 29): "Additionaly" -> "Additionally"

### Comparison to Existing Tools
- [x] I like that you mention Peter Corke's `bdsim`, but it would probably be useful to say a little bit about how PathSim is different.
- [x] I also think you should say some more about Simulink here, highlighting the obvious difference of open-source vs. proprietary.
- [x] Doing a quick search, I also found this tool called `tbcontrol` shown at https://github.com/alchemyst/Dynamics-and-Control and https://dynamics-and-control.readthedocs.io/en/latest/Simulation/Blocksim.html -- please add this, and any others you may find.

### Architecture and Design
- [x] Under "Blocks", you should probably also mention that PathSim implements subsystems to manage complexity / add hierarchy to systems.
- [x] Is there any merit to adding a new item on visualization/plotting, calling out the concept of a Scope etc.?

### PathSim Modeling Flow
- [x] It wasn't immediately obvious to me whether the block diagram visualization was something that PathSim actually offered, or whether it was just a manually created diagram to help illustrate the code. I think you should call this out more clearly in the caption.

https://github.com/openjournals/joss-reviews/issues/8158
