### 2025-05-03 #8158 @sea-bass
Hello @milanofthe -- great job on putting together a really cool software package, and the submission is in fantastic shape!

Besides general paper feedback with mostly minor comments, the main things I consider missing are documented community guidelines (required by JOSS) and some more information all around on subsystems and how to use them.

I have created 3 issues that I think should be addressed before accepting this submission:
- [ ] https://github.com/milanofthe/pathsim/issues/27
- [ ] https://github.com/milanofthe/pathsim/issues/28
- [ ] https://github.com/milanofthe/pathsim/issues/29
