### 2025-05-03 e14a17f
refactored loop dag construction in Graph to use Tarjan
