### 2025-05-03 e24397c
fixed typo in loop dag assembly after renaming
