### 2025-05-03 #8158 @danielskatz
> [@RemDelaporteMathurin](https://github.com/RemDelaporteMathurin) yes, do you mean similar to the readme, splitting up visuals and code?

I think this means in the paper.md to include the code directly, rather than an image of the code
