### 2025-05-04 969bdd5
Create CONTRIBUTING.md
First draft for community guidelines
