### 2025-05-04 5d6f3dc
added vanderpol example and vanderpol as a subsystem
