+++ docs/source/examples/thermostat.rst
@@ -1,5 +1,5 @@
 Thermostat
-----------------------
+----------
 
 In this example we have a look at a thermostat, which is a kind off discrete regulator. It switches a heating element on or of depending on defined upper and lower thresholds. 
 
