### 2025-05-04 88b26e3
added result plots to new examples
