+++ docs/source/index.rst
@@ -15,9 +15,9 @@ Key Features:
 - **Hot-swappable** blocks and solvers during simulation
 - Blocks are inherently **MIMO** (Multiple Input, Multiple Output) capable
 - Wide range of **numerical integrators** (implicit, explicit, high order, adaptive)
-- **Modular and hierarchical** modeling with (nested) subsystems
-- **Event handling** system to detect and resolve discrete events (zero-crossing detection)
-- Automatic differentiation for **fully differentiable** system simulations
+- **Modular and hierarchical** modeling with (nested) subsystems (see :ref:`ref-vanderpol-subsystem`)
+- **Event handling** system to detect and resolve discrete events (zero-crossing detection, see :ref:`ref-bouncing-ball`)
+- Automatic differentiation for **fully differentiable** system simulations (see :ref:`ref-tutorials-ad`)
 - **Extensibility** by subclassing the base `Block` class and implementing just a handful of methods
 
 The source code can be found in the `GitHub repository <https://github.com/milanofthe/pathsim>`_ and is fully open source under MIT license. Consider starring PathSim to support its development.
