### 2025-05-04 milanofthe/pathsim#27 @milanofthe
I added a [CONTRIBUTING.md](https://github.com/milanofthe/pathsim/blob/master/CONTRIBUTING.md) file to the repo.
