### 2025-05-04 milanofthe/pathsim#28 @milanofthe
Good point, I added an [example for the Subsystem](https://pathsim.readthedocs.io/en/latest/examples/vanderpol_subsystem.html#ref-vanderpol-subsystem) to the docs and crossreferences to the examples from the [index page](https://pathsim.readthedocs.io/en/latest/index.html).
