### 2025-05-04 milanofthe/pathsim#28 @sea-bass
Thank you for adding these improvements!
