### 2025-05-04 778cbb2
added readme to tests and an tests/evals dir
