+++ examples/example_vanderpol_subsystem.py
@@ -62,7 +62,7 @@ connections = [
 Sim = Simulation(
     blocks, 
     connections, 
-    Solver=GEAR52A, 
+    Solver=ESDIRK43, 
     tolerance_lte_abs=1e-5, 
     tolerance_lte_rel=1e-3, 
     tolerance_fpi=1e-8 
