### 2025-05-04 95b1b2c
update tests/README.md, removed tests/examples (will be tests/evals in the future)
