### 2025-05-04 d69cb29
added test cases to tests/evals
