+++ tests/evals/test_lorenz_system.py
@@ -1,0 +1,143 @@
+########################################################################################
+##
+##                            Testing nonlinear lorenz system
+##
+########################################################################################
+
+# IMPORTS ==============================================================================
+
+import unittest
+import numpy as np
+
+from scipy.integrate import solve_ivp
+
+from pathsim import Simulation, Connection
+from pathsim.blocks import (
+    Scope, 
+    Integrator, 
+    Constant, 
+    Adder, 
+    Amplifier, 
+    Multiplier
+    )
+
+from pathsim.solvers import (
+    RKF21, RKBS32, RKF45, RKCK54, RKDP54, RKV65, RKF78, RKDP87
+    )
+
+
+# TESTCASE =============================================================================
+
+class TestLorenzSystem(unittest.TestCase):
+
+    def setUp(self):
+
+        """Set up the system with everything thats needed 
+        for the evaluation exposed"""
+
+        # parameters 
+        sigma, rho, beta = 10, 28, 8/3
+
+        # Initial conditions
+        x0, y0, z0 = 1.0, 1.0, 1.0
+
+        # Integrators store the state variables x, y, z
+        itg_x = Integrator(x0) # dx/dt = sigma * (y - x)
+        itg_y = Integrator(y0) # dy/dt = x * (rho - z) - y
+        itg_z = Integrator(z0) # dz/dt = x * y - beta * z
+
+        # Components for dx/dt
+        amp_sigma = Amplifier(sigma)
+        add_x = Adder("+-") # Computes y - x
+
+        # Components for dy/dt
+        cns_rho = Constant(rho)
+        add_rho_z = Adder("+-") # Computes rho - z
+        mul_x_rho_z = Multiplier() # Computes x * (rho - z)
+        add_y = Adder("-+") # Computes -y + [x * (rho - z)]
+
+        # Components for dz/dt
+        mul_xy = Multiplier() # Computes x * y
+        amp_beta = Amplifier(beta) # Computes beta * z
+        add_z = Adder("+-") # Computes (x * y) - (beta * z)
+
+        # Scope for plotting
+        self.sco = Scope(labels=["x", "y", "z"])
+
+        # List of all blocks
+        blocks = [
+            itg_x, itg_y, itg_z,
+            amp_sigma, add_x,
+            cns_rho, add_rho_z, mul_x_rho_z, add_y,
+            mul_xy, amp_beta, add_z,
+            self.sco
+            ]
+
+        # Connections
+        connections = [
+            # Output signals (from integrators)
+            Connection(itg_x, add_x[1], mul_x_rho_z[0], mul_xy[0], self.sco[0]), # x -> (y-x), x*(rho-z), x*y, scope
+            Connection(itg_y, add_x[0], add_y[0], mul_xy[1], self.sco[1]),       # y -> (y-x), -y + [...], x*y, scope
+            Connection(itg_z, add_rho_z[1], amp_beta, self.sco[2]),              # z -> (rho-z), beta*z, scope
+
+            # dx/dt path: sigma * (y - x) -> itg_x
+            Connection(add_x, amp_sigma),  # (y - x) -> sigma * (y - x)
+            Connection(amp_sigma, itg_x),  # sigma * (y - x) -> integrator x
+
+            # dy/dt path: x * (rho - z) - y -> itg_y
+            Connection(cns_rho, add_rho_z[0]),     # rho -> (rho - z) input 0
+            Connection(add_rho_z, mul_x_rho_z[1]), # (rho - z) -> x * (rho - z) input 1
+            Connection(mul_x_rho_z, add_y[1]),     # x * (rho - z) -> -y + [x * (rho - z)] input 1
+            Connection(add_y, itg_y),              # x * (rho - z) - y -> integrator y
+
[…]
