### 2025-05-04 0ecabb6
adding tests for lorenz using ODE block
