### 2025-05-04 3428790
updated paper.md for JOSS, using code integration for example and fixed some typos
