### 2025-05-04 ff38941
updated paper.md, added references for tbcontrol, improved Simulink citation, added more context around subsystems
