+++ paper/paper.bib
@@ -16,6 +16,15 @@
   note         = {Source Code Repository}
 }
 
+@misc{tbcontrol,
+  author       = {Carl Sandrock},
+  title        = {A {P}ython library for solving textbook control problems},
+  howpublished = {\url{https://github.com/alchemyst/Dynamics-and-Control}},
+  month        = {May},
+  year         = {2019},
+  note         = {Source Code Repository}
+}
+
 @misc{SimuPyRepo,
   author       = {Benjamin W. l. Margolis},
   title        = {SimuPy: A framework for modeling and simulating dynamical systems},
@@ -38,16 +47,11 @@
   journal   = {Journal of Open Source Software} 
   }
 
-@software{Simulink,
-  author = {{The MathWorks, Inc.}},
-  title  = {Simulink - Simulation and model-based design},
-  url    = {https://www.mathworks.com/products/simulink.html}
-}
-
-@software{Matlab,
-  author = {{The MathWorks, Inc.}},
-  title  = {Matlab - The language of technical computing},
-  url    = {https://www.mathworks.com/products/matlab.html}
+@manual{Simulink,
+  title={{Simulink, Simulation and Model-Based Design, Version R2025a}},
+  organization={The MathWorks, Inc.},
+  address={Natick, Massachusetts},
+  year={2025}
 }
 
 @ARTICLE{virtanen2020scipy,
