### 2025-05-04 90557fc
paper.md, fixed typo
