### 2025-05-05 abcef92
refactored blocks/lti.py, now using scipy.signal to construct different kinds of transfer functions, original TransferFunction renamed to TransferFunctionPRC for pole-residue-const, adapted examples
