### 2025-05-05 02d3f61
updated docstrings
