### 2025-05-05 milanofthe/pathsim#32 @RemDelaporteMathurin
[JOSS Review] Paper feedback
This is my feedback for: openjournals/joss-reviews#8158

The only remarks I had about the paper is:

- [x] Please split the figure in two: one with the actual code block (typed instead of a screeshot). It will help readers by being able to copy paste instead of retyping everything
