### 2025-05-05 milanofthe/pathsim#29 @milanofthe
@RemDelaporteMathurin its not planned for the near future, but never say never. Maybe adding a lightweight wrapper for `graphviz` as an intermediate solution for debugging purposes is more short term.

Regrading the roadmap, there allready is one [here](https://github.com/milanofthe/pathsim/blob/master/ROADMAP.md). I can add it there and include it in the docs.
