### 2025-05-05 milanofthe/pathsim#32 @milanofthe
I also think more examples highlighting the different features (similar to the readme) would help, but then the paper is getting quite long. Not sure if thats a problem. Typical JOSS papers are relatively short.

If its no problem, I will add more examples.
