### 2025-05-05 milanofthe/pathsim#32 @RemDelaporteMathurin
I don't think this is a problem, The paper is only 3 pages long including references, I think you can add some meat ;-)
