### 2025-05-05 milanofthe/pathsim#29 @milanofthe
@sea-bass @RemDelaporteMathurin I also thought about building a frontend with ReactFlow, but I want to get the package faster and more versatile first. But longterm a GUI will probably make a lot of sense.
