### 2025-05-05 2a9fb16
Update wrapper.py to add wrapper direct access
replaced _run_wrapper method by wrapped
added classmethod dec for direct decorator access
