### 2025-05-05 b490389
Update wrapper test to valid all features with new implementation
