### 2025-05-05 2f10563
added test to check internal update of time constant in event
