### 2025-05-05 c8a9a65
Added func parameter to doc
