### 2025-05-06 0157518
adding tests for TransferFunctionNumDen and TransferFunctionZPG
