### 2025-05-06 b95f5ae
added roadmap and contribution guidelines to the docs
