+++ docs/source/roadmap.rst
@@ -1,0 +1,5 @@
+Roadmap
+=======
+
+.. include:: ../ROADMAP.md
+   :parser: myst_parser.sphinx_
