+++ CONTRIBUTING.md
@@ -1,32 +1,0 @@
-# Contributing
-Contributions to PathSim's development are very welcome! Here are some ways to help:
-
-* Improving documentation, for example with examples and explanations
-* Finding and reporting or even fixing bugs
-* Features and enhancements
-* Writing tests and validations to improve coverage
-
-For more specific ideas and inspirations, check out 
-the [roadmap](https://github.com/milanofthe/pathsim/blob/master/ROADMAP.md)!
-
-
-# Reporting Bugs
-Bugs and issues can be reported in the issue tracker using the 
-[bug report form](https://github.com/milanofthe/pathsim/issues/new). Please check the bug 
-has not already been reported or even resolved in a newer release by 
-[searching the issue tracker](https://github.com/milanofthe/pathsim/issues) before submitting a new issue!
-
-
-# Feature Requests
-Feature requests can be made in the issue tracker using the 
-[feature report form](https://github.com/milanofthe/pathsim/issues/new), 
-or by starting a [discussion](https://github.com/milanofthe/pathsim/discussions/new/choose).
-
-
-# General Support
-If you are having difficulties using PathSim or any related questions, check out the 
-[Q&A section](https://github.com/milanofthe/pathsim/discussions/categories/q-a) to see 
-if your question has already been answered. If not, feel free to start a new 
-[discussion](https://github.com/milanofthe/pathsim/discussions/new?category=q-a) 
-in the Q&A section, so everyone can benefit from the answers to your question!
-
