+++ paper/paper.bib
@@ -54,6 +54,15 @@
   year={2025}
 }
 
+@misc{ModelicaRepo,
+  title        = {Free (standard conforming) library to model mechanical (1D/3D), electrical (analog, digital, machines), magnetic, thermal, fluid, control systems and hierarchical state machines. Also numerical functions and functions for strings, files and streams are included.},
+  organization = {Modelica Association},
+  howpublished = {\url{https://github.com/modelica/ModelicaStandardLibrary}},
+  month        = {December},
+  year         = {1999},
+  note         = {Source Code Repository, created 2017}
+}
+
 @ARTICLE{virtanen2020scipy,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
