+++ paper/paper.md
@@ -213,7 +213,7 @@ Another classic dynamical system is the *Van der Pol* oscillator, which is descr
 \end{eqnarray}
 ```
 
-where the parameter $\\mu$ controls the *stiffness*. Stiffness in dynamical systems typically arises when the (local) internal time constants (eigenvalues of the Jacobian) are on vastly different scales, or when very steep gradients are encountered. This forces explicit integrators to take extremely small timesteps in order for the discretized system to be stable. Implicit solvers have larger areas of stability and are therefore a better choice for stiff problems like the *Van der Pol* system for large $\\mu$. In this example, we select the parameter $\\mu = 1000$ as a torture test do demonstrate PathSim's ability to handle severe stiffness. 
+where the parameter $\\mu$ controls the *stiffness*. Stiffness in dynamical systems typically arises when the (local) internal time constants (eigenvalues of the Jacobian) are on vastly different scales, or when very steep gradients are encountered. This forces explicit integrators to take extremely small timesteps in order for the discretized system to be stable. Implicit solvers have larger areas of stability and are therefore a better choice for stiff problems like the *Van der Pol* system for large $\\mu$. In this example, we select the parameter $\\mu = 1000$ as a torture test to demonstrate PathSim's ability to handle severe stiffness. 
 
 ![Block diagram representation of the *Van der Pol* ODE built from distinct PathSim blocks. Schematic visualization of the encapsulation into a `Subsystem`.](assets/vanderpol_g.png)
 
