### 2025-05-07 f37c45c
updated joss paper, fixed typos
