+++ paper/paper.bib
@@ -55,7 +55,7 @@
 }
 
 @misc{ModelicaRepo,
-  title        = {Free (standard conforming) library to model mechanical (1D/3D), electrical (analog, digital, machines), magnetic, thermal, fluid, control systems and hierarchical state machines. Also numerical functions and functions for strings, files and streams are included.},
+  title        = {Modelica Standard Library},
   organization = {Modelica Association},
   howpublished = {\url{https://github.com/modelica/ModelicaStandardLibrary}},
   month        = {December},
