### 2025-05-07 6889d22
fixed math display
