### 2025-05-07 milanofthe/pathsim#32 @RemDelaporteMathurin
Thanks @milanofthe , do you have a way to recompile it to a pdf?
