### 2025-05-07 milanofthe/pathsim#32 @milanofthe
@RemDelaporteMathurin usually it gets recompiled at every push, its available under the action results [here](https://github.com/milanofthe/pathsim/actions/runs/14887275832)
