### 2025-05-08 883639b
added example transfer from repo to docs as item to the roadmap
