### 2025-05-08 milanofthe/pathsim#32 @sea-bass
I know more examples / code snippets are good, but note that these papers should be pretty short!

https://joss.readthedocs.io/en/latest/paper.html
