### 2025-05-08 milanofthe/pathsim#32 @sea-bass
As for how to recompile the paper, I think on the review issue you can also type `@editorialbot generate pdf` if the latest PDF is not up to date.
