### 2025-05-08 milanofthe/pathsim#29 @sea-bass
@milanofthe let me know when you've finished making revisions to the paper and I will go through some small editorial/grammar stuff I'm already seeing. But I think the paper is in good shape from my skim of the markdown file.
