### 2025-05-08 milanofthe/pathsim#32 @milanofthe
Ok, do you think one example is ok though? Or rather just hint at the examples section?
