### 2025-05-08 d089034
Update pyproject.toml, requires Python version >= 3.8
