+++ pyproject.toml
@@ -15,7 +15,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 dependencies = ["numpy>=1.15", "matplotlib>=3.1", "scipy>=1.2", "dill>=0.3.5"]
 
 [project.optional-dependencies]
