### 2025-05-08 milanofthe/pathsim#32 @sea-bass
I think one code snippet is definitely enough.

Maybe you can do the main example with the block diagram, plots and code, and then have one more figure that just shows "a few more examples" then points to the docs.
