### 2025-05-09 f00f699
updated paper, only one example now
