+++ paper/paper.md
@@ -59,13 +59,11 @@ PathSim employs a decentralized, object-oriented design centered around three pr
 
 # PathSim Modelling Flow
 
-PathSim is a script based modelling framework, with no built in graphical user interface. Therefore it makes sense to start from the block diagram of the system to be modelled. The following sections demonstrate PathSims modelling and simulation flow. The examples cover classical dynamical systems such as the *harmonic oscillator*, the *bouncing ball* (to showcase the event handling system) and the *Van der Pol* system (as a stiff subsystem).
-
-## Harmonic Oscillator
+PathSim is a script based modelling framework with no built in graphical user interface. Therefore it makes sense to start from the block diagram of the system to be modelled. This section demonstrates PathSims modelling and simulation flow on a classical example dynamical system, the *harmonic oscillator*.
 
 The figure below shows the mechanical representation of the harmonic oscillator to the left and its block diagram to the right.
 
-![Mechanical representation of the harmonic oscillator to the left and its block diagram to the right.](assets/harmonic_oscillator_g.png)
+![Mechanical representation of the harmonic oscillator to the left and its block diagram to the right as a schematic viasualization to help with the assembly of the system im PathSim (PathSim is purely script based)](assets/harmonic_oscillator_g.png)
 
 Translating it to PathSim involves importing the blocks from the block library, instantiating them with their correct parameters, connecting them with the `Connection` object and passing everything to the `Simulation`.
 
@@ -115,175 +113,7 @@ The `Scope` block enables fast plotting of the simulation results as well as the
 
 ![Simulation results plotted from the PathSim `Scope` for the harmonic oscillator example.](assets/harmonic_oscillator_result_g.png)
 
-
-## Bouncing Ball
-
-The classical example for demonstrating event handling is the bouncing ball. The ball accelerates downward due to gravity and onece it hits the floor, it bounces back. The dynamics of this system are discontinuous due to the jump in velocity (sign change) exactly at the bounce. This is called a discrete event, specifically a zero-crossing event. Systems like this require the location and resolution of discrete events within the timestep and are not easily solved with standard numerical ODE solvers. **PathSim** implements an event management system, that can track the system state and locate and resolve specified events. 
-
-![Schematic of the bouncing ball dynamics to the left and the corresponding block diagram with two distinct integrators and the event handler to the righ.](assets/bouncing_ball_g.png)
-
-The following code block defines the bouncing ball system in PathSim, utilizing the `ZeroCrossing` event manager to track the height of the ball and trigger the jump in velocity.
-
-```python
-from pathsim import Simulation, Connection
-from pathsim.blocks import Integrator, Constant, Scope
-from pathsim.events import ZeroCrossing #event manager
-from pathsim.solvers import RKBS32 #adaptive solver
-
-#elasticity of bounce and gravitational acceleration
-b, g = 0.9, 9.81
-
-#initial conditions
-x0, v0 = 1, 5
-
-#blocks that define the system
-Ix = Integrator(x0)     # v -> x
-Iv = Integrator(v0)     # a -> v 
-Cn = Constant(-g)       # gravitational acceleration
-Sc = Scope(labels=["x", "v"])
-
-blocks = [Ix, Iv, Cn, Sc]
-
-#the connections between the blocks
-connections = [
[…]
