### 2025-05-09 milanofthe/pathsim#32 @milanofthe
I updated the paper to only feature the harmonic oscillator example and added a small hint to more examples in the repo and docs at the end. Im not quite sure I get what you mean by *one more figure that just shows "a few more examples"*.
