### 2025-05-09 97029d8
simplified volterralotka example
