+++ examples/examples_odes/example_volterralotka.py
@@ -23,43 +23,43 @@ delta = 0.5  # predator efficiency
 gamma = 1.2  # death rate of predators
 
 
+#blocks that define the system
 i_pred = Integrator(10)
 i_prey = Integrator(5)
 
-c_alp = Constant(alpha)
-c_gma = Constant(gamma)
-
+a_alp = Amplifier(alpha)
+a_gma = Amplifier(gamma)
 a_bet = Amplifier(beta)
 a_del = Amplifier(delta)
 
 p_pred = Adder("-+")
 p_prey = Adder("+-")
 
-m_pred = Multiplier()
-m_prey = Multiplier()
+m_pp = Multiplier()
 
 sco = Scope(labels=["predator population", "prey population"])
 
 blocks = [
-    i_pred, i_prey, c_alp, c_gma, 
+    i_pred, i_prey, a_alp, a_gma, 
     a_bet, a_del, p_pred, p_prey, 
-    m_pred, m_prey, sco
+    m_pp, sco
     ]
 
+#the connections between the blocks
 connections = [
-    Connection(i_pred, m_pred[0], a_del, sco[0]),
-    Connection(i_prey, m_prey[0], a_bet, sco[1]),
+    Connection(i_pred, m_pp[0], a_alp, sco[0]),
+    Connection(i_prey, m_pp[1], a_gma, sco[1]),
     Connection(a_del, p_prey[0]),
-    Connection(c_gma, p_prey[1]),
+    Connection(a_gma, p_prey[1]),
     Connection(a_bet, p_pred[0]),
-    Connection(c_alp, p_pred[1]),
-    Connection(p_pred, m_pred[1]),
-    Connection(p_prey, m_prey[1]),
-    Connection(m_pred, i_pred),
-    Connection(m_prey, i_prey)
+    Connection(a_alp, p_pred[1]),
+    Connection(m_pp, a_del, a_bet),
+    Connection(p_pred, i_pred),
+    Connection(p_prey, i_prey)
     ]
 
 
+#initialize the simulation with everything
 Sim = Simulation(
     blocks, 
     connections, 
