### 2025-05-09 ce8fa05
added volterra lotka system to tests/evals
