### 2025-05-09 5bb2f61
adding vanderpol system as a subsystem to tests/evals
