### 2025-05-11 milanofthe/pathsim#32 @sea-bass
> Im not quite sure I get what you mean by _one more figure that just shows "a few more examples"_.

I mean that there can be another figure that shows block diagrams and plots for a couple more representative examples, without the code snippet to take up space. Just as a visual way to say "yes, you can simulate other things".
