### 2025-05-13 5ad1ac3
adding example for algebraic signal chain
