### 2025-05-13 a0c0e48
adding test case for algebraic system to tests/evals
