### 2025-05-13 dc91169
paper figure captions
