### 2025-05-13 a4bad6c
paper figure captions
