### 2025-05-13 milanofthe/pathsim#29 @milanofthe
@sea-bass I added the two examples as figures without the code (is this what you had in mind?) to the paper and also made some adjustments to the other figures. The new figures are primarily described trough the captions.

Other than that, I am done with the paper.
