### 2025-05-13 milanofthe/pathsim#29 @sea-bass
Great! Just small grammar things:

* "Its fully open under MIT-License" -> It is fully open...
* "This section demonstrates PathSims modelling and simulation flow" -> PathSim's modelling...

So yeah, looks great and I consider this done, but please fix those 2 small things :)
