### 2025-05-13 milanofthe/pathsim#32 @danielskatz
@RemDelaporteMathurin - are you ok with this issue being closed?
