### 2025-05-13 #8158 @RemDelaporteMathurin
I also recommend this submission for publication in JOSS 🎉

Good job @milanofthe
