+++ paper/paper.bib
@@ -77,8 +77,8 @@
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in {P}ython}},
+  title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in {P}ython},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
