### 2025-05-13 #8158 @danielskatz
👋 @milanofthe - I've suggested some minor changes in https://github.com/milanofthe/pathsim/pull/44

Also, I notice there is no Acknowledgments sections in the paper. Did you want to add one, to mention any funding sources or anything else?
