### 2025-05-13 #8158 @milanofthe
@danielskatz thanks! I followed through with your points:

* the current release (from master branch) is [v0.6.9](https://github.com/milanofthe/pathsim/releases/tag/v0.6.9)
* the repository is archived in Zenodo under this DOI: [10.5281/zenodo.15367933](https://doi.org/10.5281/zenodo.14992799) (title, author and ORCID matches)
